Port PathsTest
